# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Baxi Brothers ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  productCount: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice: number | null;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  imageUrl: string | null;
  images: string[];
  inStock: boolean;
  stockCount: number;
  featured: boolean;
  /** @nullable */
  material: string | null;
  /** @nullable */
  dimensions: string | null;
  /** @nullable */
  careInstructions: string | null;
  tags: string[];
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  inStock: boolean;
  stockCount: number;
  featured: boolean;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  careInstructions?: string | null;
  tags?: string[];
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
  description?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId?: number;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  inStock?: boolean;
  stockCount?: number;
  featured?: boolean;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  careInstructions?: string | null;
  tags?: string[];
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  limit: number;
  offset: number;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  imageUrl: string | null;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: string;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  total: number;
  status: OrderStatus;
  items: OrderItem[];
  createdAt: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface DashboardStats {
  totalProducts: number;
  totalCategories: number;
  inStockProducts: number;
  outOfStockProducts: number;
  featuredProducts: number;
}

export interface CategoryCount {
  categoryId: number;
  categoryName: string;
  count: number;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
inStock?: boolean;
featured?: boolean;
limit?: number;
offset?: number;
};