# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Baxi Brothers ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the signed-in customer's orders
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order for the signed-in customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
        "401":
          description: Authentication required
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one of the signed-in customer's orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Customer order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
        "404":
          description: Not found
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all customer orders for the admin panel
      responses:
        "200":
          description: All customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
  /admin/orders/{id}:
    get:
      operationId: getAdminOrder
      tags: [admin]
      summary: Get a complete order for the admin panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Complete customer order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
        "404":
          description: Not found
    patch:
      operationId: updateAdminOrder
      tags: [admin]
      summary: Update an order's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Authentication required
        "404":
          description: Not found
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, description, imageUrl, productCount, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: number
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name, slug, description]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
...
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: number
        quantity:
          type: number
          minimum: 1
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, imageUrl]
      properties:
        id:
          type: number
        productId:
          type: number
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        imageUrl:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [customerName, email, phone, address, city, items]
      properties:
        customerName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    Order:
      type: object
      required: [id, userId, customerName, email, phone, address, city, total, status, items, createdAt]
      properties:
        id:
          type: number
        userId:
          type: string
        customerName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
...