# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Baxi Brothers ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
}