# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Baxi Brothers ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: string;
  customerName: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  total: number;
  status: OrderStatus;
  items: OrderItem[];
  createdAt: string;
}