# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Baxi Brothers ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  imageUrl: string | null;
}